Commit 5 for part 3 of assignment, took out decimal for withdrawal when printing results. Kept decimal for new balance, as new balance can have decimal parts (for instance, if they withdraw 50 and not 100).

+++ src/Module2_Homework3.java
@@ -44,7 +44,7 @@ public class Module2_Homework3 {
             WithdrawalReturn = bankwithdrawal(balance, withdrawal, commissions); // calling method
 
             // printing results
-            System.out.println((name + " " + WithdrawalReturn[0] + " " + WithdrawalReturn[1] ));
+            System.out.println((name + " " + (int) WithdrawalReturn[0] + " " + WithdrawalReturn[1] ));
         } else {
             System.out.println(name + "  NO");
         }
